Separated integration tests in different files
